requirements.txt
Django>=1.5
djangorestframework>=2.4.5
